consumer/requirements.txt
redis==2.10.6
kafka-python==2.0.1
requests==2.28.2
pandas==2.2.2
python-dotenv==1.0.1
ipython==8.18.1
pandas==2.2.2
grpcio-tools==1.59.3
grpcio==1.59.3
loguru==0.7.2
asyncio==3.4.3
apache-flink==1.18.0
redis==2.10.6
starlette==0.27.0
